Simplify the encrypted? method
Make this easier to read by using built in Hash methods and also avoid creating new Hashes on each operation. The real reason here is that the previous method of calculating if an encrypted data bag was encrypted involved checking each value to see if it was encrypted and then comparing all the values that were encrypted to the original hash. At the point that we find the first unencrypted value we can bail out and avoid checking the whole hash. This could potentially save a bit of time for very large data bags that aren't encrypted.

Signed-off-by: Tim Smith <tsmith@chef.io>